+++ HISTORY.md
@@ -12,6 +12,12 @@
 - Renamed event_det_port und data_fetch_port
 - Simplify log setup (no separated calls for with or without onscreen)
 
+# 4.0.17
+
+- Fixed MOVE_TO events in watchdog event detector
+- Fixed hangin of datadispatcher
+- Fixed missing events in watchdog event detector
+
 # 4.0.16
 
 - Fixed control signal reaction during data handling